requirements.txt
torch
torchvision
flask
opencv-python
numpy
tqdm
Pillow